refactor(dashboard): 重构文章监听和配置生成逻辑
